[#156852443] Write test for MealController getMeals method
