+++ server/controllers/controller.js
@@ -1,5 +1,5 @@
 class Controller {
-
+	
 }
 
 export default Controller;
